isolated authentication layer
