+++ sql_app/main.py
@@ -1,15 +1,14 @@
 from fastapi import Depends, FastAPI, HTTPException, Response, Request
 from starlette.status import HTTP_204_NO_CONTENT, HTTP_202_ACCEPTED, HTTP_201_CREATED, HTTP_200_OK, HTTP_401_UNAUTHORIZED
 from sqlalchemy.orm import Session
+from sql_app.auth import ACCESS_TOKEN_EXPIRE_MINUTES, authenticate_user, create_access_token, get_current_active_user, fake_users_db
 from . import db_queries, models, schemas
 from .database import SessionLocal, engine
 from logging.config import dictConfig
 import logging
 from .log_config import logging_schema_api
 from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestFormStrict
-from jose import JWTError, jwt
-from passlib.context import CryptContext
-from datetime import datetime, timedelta
+from datetime import timedelta
 
 import uvicorn
 
@@ -141,66 +140,6 @@ def get_genres(response: Response, request: Request, db: Session = Depends(get_d
     return db_genres
 
 
-def verify_password(plain_password, hashed_password):
-    return pwd_context.verify(plain_password, hashed_password)
-
-
-def get_password_hash(password):
-    return pwd_context.hash(password)
-
-
-def get_user(db, username: str):
-    if username in db:
-        user_dict = db[username]
-        return schemas.UserInDB(**user_dict)
-
-
-def authenticate_user(fake_db, username: str, password: str):
-    user = get_user(fake_db, username)
-    if not user:
-        return False
-    if not verify_password(password, user.hashed_password):
-        return False
-    return user
-
-
-def create_access_token(data: dict, expires_delta: timedelta | None = None):
-    to_encode = data.copy()
-    if expires_delta:
-        expire = datetime.utcnow() + expires_delta
-    else:
-        expire = datetime.utcnow() + timedelta(minutes=15)
-    to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
-
-
-async def get_current_user(token: str = Depends(oauth2_scheme)):
-    credentials_exception = HTTPException(
-        status_code=HTTP_401_UNAUTHORIZED,
-        detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username: str = payload.get("sub")
-        if username is None:
-            raise credentials_exception
-        token_data = schemas.TokenData(username=username)
-    except JWTError:
-        raise credentials_exception
-    user = get_user(fake_users_db, username=token_data.username)
-    if user is None:
-        raise credentials_exception
-    return user
-
-
-async def get_current_active_user(current_user: schemas.User = Depends(get_current_user)):
-    if not current_user.is_active:
-        raise HTTPException(status_code=400, detail="Inactive user")
-    return current_user
-
-
 @app.post("/api/v1/token", response_model=schemas.Token)
 async def login_for_access_token(form_data: OAuth2PasswordRequestFormStrict = Depends()):
     user = authenticate_user(
@@ -221,23 +160,6 @@ async def login_for_access_token(form_data: OAuth2PasswordRequestFormStrict = De
 @app.get("/api/v1/users/me", response_model=schemas.User)
 async def read_users_me(current_user: schemas.User = Depends(get_current_active_user)):
     return current_user
-
-
-SECRET_KEY = "748e5f50fbd00e5dc7ae583e7a5bf27d001f43a17970f6c9796d67b3294da59b"
-ALGORITHM = "HS256"
-ACCESS_TOKEN_EXPIRE_MINUTES = 120
-
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-fake_users_db = {
-    "admin": {
-        "username": "admin",
-        "full_name": "John Doe",
-        "email": "admin@admin.com",
-        "hashed_password": "$2b$12$EixZaYVK1fsbw1ZfbX3OXePaWxn96p36WQoeG6Lruj3vjPGga31lW",
-        "is_active": True,
-    }
-}
 
 
 @app.delete("/api/v1/movies/{movie_id}", status_code=HTTP_202_ACCEPTED, response_model=schemas.Movie)
